requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
blinker==1.8.2
build==1.2.2.post1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
cohere==5.13.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.3
ctransformers==0.2.27
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
durationpy==0.9
emoji==2.14.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0
fastapi==0.115.2
fastavro==1.10.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gpt4all==2.8.2
greenlet==3.0.3
groq==0.15.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
kubernetes==31.0.0
langchain==0.3.13
langchain-chroma==0.1.4
langchain-cohere==0.3.4
langchain-community==0.3.13
langchain-core==0.3.30
langchain-experimental==0.3.4
langchain-google-genai==2.0.7
langchain-groq==0.2.3
langchain-huggingface==0.1.0
langchain-ollama==0.2.2
langchain-text-splitters==0.3.4
langdetect==1.0.9
langsmith==0.1.136
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.4
onnxruntime==1.19.2
openai==1.42.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
posthog==3.7.0
proto-plus==1.25.0
protobuf==5.29.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.2.0
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.1
PyYAML==6.0.2
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.7.24
replicate==1.0.4
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.3.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
starlette==0.40.0
streamlit==1.37.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1
torchaudio==2.4.1+cu118
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.46.1
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.16.0
unstructured-client==0.26.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.32.0
watchdog==4.0.2
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.2
